Remade the whole system
Because I dident like al the if and else statements cluttering
everywhere and because I also didn’t like the way the action object was
made, I remade the whole thing. I striped the action object into a nlp
object which only has the essentials. I then made all msg got to the
msgRouter, it routes messages to the correct plugin and/or sends
message to user if there is no plugin. tldr it takes care of all the
msg…and loads all the plugins too.